S.I.E.R.I.A/requirements.txt
colorama
requests
scapy
python-nmap
paramiko
cryptography
stem